+++ src/render/primitives/Sphere.cpp
@@ -118,10 +118,10 @@ RayPacketIntersection4 Sphere::intersectPacket(const Ray4& rays, render::State& 
 #endif
 }
 
-PrimitivePacketHit4 Sphere::intersectPacketHits(const Ray4& rays,
-                                                const PrimitivePacketState4& states) const {
-  PrimitivePacketHit4 result;
-  for (std::size_t lane = 0; lane != Ray4::lanes; ++lane) {
+template<typename Packet, typename StateArray, typename Result>
+Result Sphere::intersectPacketHitsFor(const Packet& rays, const StateArray& states) const {
+  Result result;
+  for (std::size_t lane = 0; lane != Packet::lanes; ++lane) {
     State fallbackState;
     State& state = states[lane] ? *states[lane] : fallbackState;
     RAYTRACER_STATS_INC(raySphereIntersect);
@@ -152,6 +152,16 @@ PrimitivePacketHit4 Sphere::intersectPacketHits(const Ray4& rays,
     state.hit(this, "Sphere");
   }
   return result;
+}
+
+PrimitivePacketHit4 Sphere::intersectPacketHits(const Ray4& rays,
+                                                const PrimitivePacketState4& states) const {
+  return intersectPacketHitsFor<Ray4, PrimitivePacketState4, PrimitivePacketHit4>(rays, states);
+}
+
+PrimitivePacketHit8 Sphere::intersectPacketHits(const Ray8& rays,
+                                                const PrimitivePacketState8& states) const {
+  return intersectPacketHitsFor<Ray8, PrimitivePacketState8, PrimitivePacketHit8>(rays, states);
 }
 
 bool Sphere::intersects(const Rayd& ray, render::State& state) const {